requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
bcrypt
python-multipart
pyjwt
boto3
stripe
mangum
moment